requirements.txt
streamlit
numpy
pandas
matplotlib
opencv-python-headless
scikit-image
Pillow
streamlit-cropper